@@ -5,20 +5,13 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace BlackHole._360.Api.Controllers;
 
-public class UsersController : BaseController
+public class UsersController(UserService userService) : BaseController()
 {
-    private readonly UserService _userService;
-
-    public UsersController(UserService userService) : base() 
-    {
-        _userService = userService;
-    }
-
     [HttpGet("{id}")]
     public async Task<ActionResult<UserDto>> IndexAsync(Guid id, CancellationToken cancellationToken = default)
-        => Ok(await _userService.GetAsync(id, cancellationToken));
+        => Ok(await userService.GetAsync(id, cancellationToken));
 
     [HttpGet]
     public async Task<ActionResult<IEnumerable<UserDto>>> IndexAsync([FromQuery]string search, [FromQuery]int offset, [FromQuery]int count, CancellationToken cancellationToken)
-        => Ok(await _userService.GetAsync(search, offset, count, cancellationToken));
+        => Ok(await userService.GetAsync(search, offset, count, cancellationToken));
 }
