@@ -3,10 +3,8 @@ using Microsoft.EntityFrameworkCore;
 using System.Reflection;
 
 namespace BlackHole._360.DataAccess;
-public class BlackHoleContext : DbContext
+public class BlackHoleContext(DbContextOptions<BlackHoleContext> options) : DbContext(options)
 {
-    public BlackHoleContext(DbContextOptions<BlackHoleContext> options) : base(options) { }
-
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
         modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
